deepfake-detection-app/requirements.txt
streamlit==1.45.1
pillow==10.1.0
numpy==1.26.3
opencv-python==4.9.0.80
tf-nightly>=2.16.1 